Do not delete Imported Service Records
We were getting away with this so far in our e2e test because contour
wasn't present on the workload cluster, and therefore the Imported
ServiceRecords were failing to delete because the publisher wasn't able
to resolve the Contour HTTPProxy object.

Adding the contour deployment means that the publisher was now
erroneously deleting the Imported Service Records.

The publisher should only delete orphaned exported Service Records, and
not imported ones.

Signed-off-by: Jamie Monserrate <monserratej@vmware.com>