Delete ServiceRecords that are no longer allowed
Modifies registryclient to delete ServiceRecords with FQDNs that are no
longer allowed when reconciling ServiceRecords.

This behavior relies on a part of the Hamlet protocol that sends all
known Federated Services during the handshake. Because a change in the
RemoteRegistry resource triggers a redial, this re-sync of all
Federated Services will occur on any change to the AllowedDomains list
in the RemoteRegistry resource.

There is still a case where the registryclient will not clean up
a ServiceRecord that has an FQDN that does not match the AllowedDomains
list. This case occurs when the registryclient controller does not
reconcile a particular ServiceRecord, and this can only occur if the
FederatedService was deleted on the server side without the
registryclient being notified. Handling of orphans in this situation can
be done in a separate effort.