Fix data race in registryclient unit tests
For some reason, using the Kubernetes generated fakes causes a data race
between structs that exist inside the unit tests and the structs used
inside the controller sync logic. It's not limited to just the
allowedDomains field, so the controller will now deep copy the entire
RemoteRegistry struct upon retrieval from the API.

This does not appear to be a concern for runtime code, so perhaps in the
future there is a better way to fix this data race issue in the unit
tests.