Extend e2e test curl timeout
There appears to be some period of time after `make e2e-up` finishes
where connectivity cannot be established until the deployment has had
some time to warm up. Bumping the timeout in the e2e test will give the
deployment more time to warm up before testing for connectivity.

We think it was green in CI until now because the test is/was just under
the 60 second timeout. This test was passing locally, but in 50 seconds
(v/s the 60 in CI).

Co-authored-by: Edwin Xie <exie@vmware.com>