merge check-license into test workflow
We wanted to have a workflow that would fan-in on completion of Test and
Check License to run Publish Docker Image. However it appears that
GitHub actions don't support this behavior.

Check License is quick enough to be combined with the Test stage. This
is the simplest way to get the workflow we desired.

 #62

Co-authored-by: Tyler Schultz <tschultz@vmware.com>