Hostnames use the namespace of the Cluster resource
The namespace in the management cluster containing the workload Cluster
is what goes into the DNS hostname.  It happens to be the same namespace
containing the GatewayDNS today, but that's conincidental.
They may not be the same in the future. This should make it easier to
support cross-namespace reconcilation in the future.

Co-authored-by: Tyler Schultz <tschultz@vmware.com>